pyproject.toml
[project]
name = "file-rename"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyexiftool>=0.5.6",
]
